+++ MarkdownNotesApp/ClientApp/react-ts-notes/src/@types/notes.d.ts
@@ -1,0 +1,23 @@
+export type Note = {
+    noteId: string
+    title: string
+    markdown: string
+    tags: Tag[]
+}
+
+export type NoteDto = {
+    title: string
+    markdown: string
+    tagCreateDtos: TagDto[]
+}
+
+export type Tag = {
+    tagId: string
+    label: string
+}
+
+export type TagDto = {
+    label: string
+}
+
+
