+++ MarkdownNotesApp/ClientApp/react-ts-notes/src/components/NoteForm.tsx
@@ -10,10 +10,10 @@ type NoteFormProps = {
     title?: string
     markdown?: string
     tags?: TagDto[]
-} 
+}
 //NoteData is optional this way and newNote and editNote can both use this component
 
-export function NoteForm({ onSubmit, availableTags, title="", markdown="",tags=[] } : NoteFormProps) {
+export function NoteForm({ onSubmit, availableTags, title = "", markdown = "", tags = [] }: NoteFormProps) {
     const titleRef = useRef<HTMLInputElement>(null)
     const markdownRef = useRef<HTMLTextAreaElement>(null)
     const [selectedTags, setSelectedTags] = useState<TagDto[]>(tags)
@@ -40,35 +40,35 @@ export function NoteForm({ onSubmit, availableTags, title="", markdown="",tags=[
                     <Col>
                         <Form.Group controlId="title">
                             <Form.Label>Title</Form.Label>
-                            <Form.Control required ref={titleRef} defaultValue={title} />
+                            <Form.Control className="input-field" required ref={titleRef} defaultValue={title} />
                         </Form.Group>
                     </Col>
                     <Col>
                         <Form.Group controlId="tags">
                             <Form.Label>Tags</Form.Label>
-                            <CreatableReactSelect 
-                            options={availableTags.map(tag => {
-                                return { label: tag.label, value: tag.label }
-                            })}
-                            onCreateOption={label => {
-                                const newTag = { label: label }
-                                setSelectedTags(prev => [...prev, newTag])
-                            }}
-                            value={selectedTags.map(tag => {
-                                return { label: tag.label, value: tag.label }
-                            })} 
-                            onChange={tags => {
-                                setSelectedTags(tags.map(tag => {
-                                    return { label: tag.label }
-                                }))
-                            }}
-                            isMulti />
+                            <CreatableReactSelect
+                                options={availableTags.map(tag => {
+                                    return { label: tag.label, value: tag.label }
+                                })}
+                                onCreateOption={label => {
+                                    const newTag = { label: label }
+                                    setSelectedTags(prev => [...prev, newTag])
+                                }}
+                                value={selectedTags.map(tag => {
+                                    return { label: tag.label, value: tag.label }
+                                })}
+                                onChange={tags => {
+                                    setSelectedTags(tags.map(tag => {
+                                        return { label: tag.label }
+                                    }))
+                                }}
+                                isMulti />
                         </Form.Group>
                     </Col>
                 </Row>
                 <Form.Group controlId="markdown">
                     <Form.Label>Body</Form.Label>
-                    <Form.Control required as="textarea" ref={markdownRef} rows={15} defaultValue={markdown} />
+                    <Form.Control className="input-field" required as="textarea" ref={markdownRef} rows={15} defaultValue={markdown} />
                 </Form.Group>
                 <Stack direction="horizontal" gap={2} className="justify-content-end">
                     <Button type="submit" variant="primary">Save</Button>
